Refactor yolo converter into a module for testing
